Update the specification of the endpoints
I'll follow the JSON:API specification and tried to applied some best
practices for REST APIs.

The endpoint name is pluralized.

I aim to define an API of Level 3 with hypermedia links, according to
the Richardson Maturity Model.

For that reason I've included some links on every endpoint, so the
client could know what to do next.

References:

* https://blog.restcase.com/7-rules-for-rest-api-uri-design/
* https://martinfowler.com/articles/richardsonMaturityModel.html
* https://jsonapi.org/
* https://swagger.io/specification/v3/
* https://swagger.io/docs/specification/links/#parameters
* https://swagger.io/docs/specification/describing-parameters/
* https://stackoverflow.com/questions/50204588/how-to-define-uuid-property-in-json-schema-and-open-api-oas
* https://www.base64decode.org/