Add endpoint to create new safeboxes
The endpoint is configured with the minimum code required to know that
all the pieces are working as expected.

Nothing has been implemented at this moment, so it is currently
returning a status 200 and an empty body.

Following the approach suggested by the people from Codely, I'll use a
dedicated directory, apps, for all the possible code using the bounded
contexts instead of using the default app directory.

I rather prefer having more flexibility with that code, mainly due to
the way the code is autoloaded within the default directory.

The endpoints are namespaced with the version of the API.

Regarding the naming of the controllers, I'm using a more semantic name,
as I'll use a controller for each endpoint, instead of adding all
endpoints in a single controller.

Also suggested by Codely, the ID of the safebox will be provided from the
client instead of being autogenerated in the database (check video
below). The main difference is in the endpoint itself, as they prefer to
use the endpoint to update the resource because the ID is already known.
I can see it, but here I stick with the classic endpoint (POST method).

References:

* https://guides.rubyonrails.org/autoloading_and_reloading_constants.html
* https://github.com/fxn/zeitwerk
* https://swagger.io/docs/specification/links/
* https://github.com/rswag/rswag
* https://github.com/rswag/rswag/issues/685
* https://github.com/rswag/rswag/blob/master/rswag-specs/lib/rswag/specs/request_factory.rb#L204
* https://github.com/rswag/rswag/issues/359
* https://www.rubydoc.info/gems/rubocop-rspec/RuboCop/Cop/RSpec/NestedGroups
* https://www.rubydoc.info/gems/rubocop-rspec/RuboCop/Cop/RSpec/DescribeClass
* https://www.youtube.com/watch?v=wR5b0OhbUyw
* https://www.youtube.com/watch?v=P3bpYv9vdic