Add use case to create new safeboxes
These changes are again highly inspired by Codely.

Uses dry and some of the gems of its ecosystem to enable auto dependency
injection in Rails.

Every use case will receive a repository injected in its constructor, that
must implement a defined interface. A custom exception would be raised
in case the repository does not implement the interface.

The use case will return a result object either with defined objects on
success or with found errors.

Besides, an input is provided to an instance of a use case, that will
validate provided data or parameters from the controller or any other
entrypoint of the application.

That logic regarding use cases and inputs is highly inspired (read
stolen) from Upgrow by Shopify. More details in a link below.

A new aggregate root has been defined to represent a new safebox, that
has an ID, a name and a password. Later on, this aggregate root won't be
used, because what we have in the database is the password digest, that
will be used to validate provided password, but won't be exposed. So
clearly two different aggregate roots representing the same resource in
different states.

Every attribute of that aggregate root is a value object that itself
only has a value attribute. With dry-types is quite easy defining
validations for every value object.

The validations added for name and password are added for the sake of
showing how validations are handled in the application.

In the case of the password, it's limited to 72 bytes due to a known
limitation by bcrypt, the gem used to encrypt passwords.

From the controller we are able to validate all given data at once with
the schema validator created previously, but here a custom exception is
raised at the very moment an error is found.

The tests for the use case acts as unit tests in this application. Using
mocks we avoid creating records in the database.

The next step will be creating integration tests for the repository
injected in the use case, where some records will be created in the
database.

References:

* https://dry-rb.org/
* https://dry-rb.org/gems/dry-rails
* https://dry-rb.org/gems/dry-system
* https://dry-rb.org/gems/dry-types
* https://dry-rb.org/gems/dry-struct
* https://dry-rb.org/gems/dry-auto_inject
* https://dry-rb.org/gems/dry-container
* https://github.com/dry-rb/dry-rails/blob/main/lib/dry/rails/container.rb
* https://github.com/dry-rb/dry-system/blob/main/lib/dry/system/container.rb
* https://github.com/dry-rb/dry-system/blob/main/lib/dry/system/provider_registrar.rb
* https://github.com/backpackerhh/upgrow-docs/blob/main/guide/6-the-missing-pieces.md
* https://docs.ruby-lang.org/en/3.2/Random/Formatter.html#method-i-alphanumeric
* https://api.rubyonrails.org/v7.1.3/classes/ActiveModel/SecurePassword/ClassMethods.html
* https://api.rubyonrails.org/classes/ActiveModel/SecurePassword.html
* https://docs.rubocop.org/rubocop/cops_style.html#styleifunlessmodifier
* https://docs.rubocop.org/rubocop/cops_style.html#stylenegatedif
* https://docs.rubocop.org/rubocop-rspec/cops_rspec.html#rspecmessagespies
* https://docs.rubocop.org/rubocop-rspec/cops_rspec.html#rspecexamplelength
* https://docs.rubocop.org/rubocop-rspec/cops_rspec.html#rspecverifieddoublereference
* https://docs.rubocop.org/rubocop-rspec/cops_rspec.html#rspecverifieddoubles