Add logic to create new safeboxes
All possible responses from the server are now covered.

Although is not expected to be used, the response after creating a new
safebox will the minimum data required to be JSON:API compliant.

Request specs act as acceptance test here and with these changes, the
outside-in cycle has been completed.

The strict mode has been disabled in schema validation because that mode
marks all properties as required.

The Accept header is required with the value "application/vnd.api+json" in
every request. Otherwise the server will return a 406 response.

The Content-Type header is required with the value
"application/vnd.api+json" in every request. Otherwise the server will
return a 415 response.

All errors are nicely handled so all of them return a response with the
same format.

References:

* https://github.com/jsonapi-serializer/jsonapi-serializer
* https://jsonapi.org/format/#document-structure
* https://jsonapi.org/format/#content-negotiation
* https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Accept
* https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Type
* https://developer.mozilla.org/en-US/docs/Web/HTTP/Content_negotiation
* https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/201
* https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/406
* https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/409
* https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/415
* https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/422
* https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/500
* https://www.honeybadger.io/blog/understanding-the-ruby-exception-hierarchy/
* https://gist.github.com/mlanett/a31c340b132ddefa9cca
* https://github.com/rswag/rswag/blob/master/rswag-specs/lib/rswag/specs/response_validator.rb
* https://github.com/voxpupuli/json-schema?tab=readme-ov-file#advanced-options
* https://github.com/voxpupuli/json-schema/blob/master/lib/json-schema/validator.rb
* https://ruby-doc.org/core-2.4.1/Regexp.html#method-c-new