Add use case to open safeboxes
This code applies the same logic that was already applied for the use
case that allows to create new safeboxes. Refer to the related commit
for more details.

Safeboxes are locked after 3 failing attempts to open them. That could
happen if the encoded password provided in a "X-SafeIsh-Password" header
was not defined or was invalid.

These unit tests are currently in green, mocking the repository received
in the constructor.

Once again, the next step will be creating integration tests for the
repository.

References:

* https://docs.rubocop.org/rubocop/cops_metrics.html#metricsabcsize
* https://docs.rubocop.org/rubocop-rails/cops_rails.html#railsdynamicfindby