Allow to generate tokens in repository to open safeboxes
The safebox ID and the password added on creation are used to open a
given safebox.

The token returned once the safebox password has been validated is
generated using a functionality by Rails, and will be valid for 3
minutes.

Every time an attempt to open a safebox is made with an invalid
password, an attribute on the safeboxes table will keep track of that.
After 3 failed attempts the sabebox will be locked.

In the same way that are two different entities for new safeboxes and
for safeboxes loaded from the database, now there are two different
entity factories for each one of them.

The equality of an aggregate root will use from now on not only the ID
but the object class as well.

The integration tests for the repository are currently in green.

References:

* https://edgeapi.rubyonrails.org/classes/ActiveRecord/TokenFor/ClassMethods.html#method-i-generates_token_for
* https://blog.saeloun.com/2023/11/14/rails-7-1-introduces-active-record-generate-token-for/
* https://github.com/rails/rails/pull/44189
* https://ruby-doc.org/stdlib-2.5.3/libdoc/base64/rdoc/Base64.html#method-i-decode64
* https://github.com/rails/rails/blob/main/activemodel/lib/active_model/secure_password.rb
* https://docs.rubocop.org/rubocop-rails/cops_rails.html#railsskipsmodelvalidations