Add missing logic required to open a safebox
All possible responses from the server are now covered.

The acceptance tests are now in green.

Notice that in case the password is either not provided or is invalid, a
not authorized response (401) will be returned.

In case of success, an ok/success response (200) will be returned.

Regarding the endpoint, a post method is used because it is not
idempotent. Every call to that endpoint would return a different token
to open the safebox.

Arguably, a created response (201) could be used as well.

References:

* https://docs.rubocop.org/rubocop/cops_metrics.html#metricsmethodlength