Add use case to list safebox items
This code applies the same logic than previous use cases. Refer to the
related commits for more details.

A token is used for granting access to the contents of the safebox.

A domain service is used for lazy loading the items associated to a
given safebox.

These unit tests are currently in green, mocking the repository received
in the constructor.

Once again, the next step will be creating integration tests for the
repository.

References:

* https://api.rubyonrails.org/classes/ActiveModel/AttributeMethods/ClassMethods.html#method-i-alias_attribute