Add missing logic required to listing the contents of a safebox
All possible responses from the server are now covered.

The acceptance tests are now in green.

Notice that in case the token is not provided, is invalid or has expired, a
not authorized response (401) will be returned.

Besides, WWW-Authenticate header will be returned in the response as well.

A workaround to serialize the correct ID of a relationship has been used
for safebox items in both the serializer and the entity. More details in
the GitHub issue linked below.

References:

* https://swagger.io/docs/specification/authentication/bearer-authentication/
* https://swagger.io/docs/specification/describing-responses/
* https://developer.mozilla.org/en-US/docs/Web/HTTP/Authentication
* https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Authorization
* https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/401
* https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/WWW-Authenticate
* https://stackoverflow.com/questions/44323531/how-to-get-bearer-token-passed-through-header-in-rails
* https://datatracker.ietf.org/doc/html/rfc6750
* https://datatracker.ietf.org/doc/html/rfc2617
* https://swagger.io/docs/specification/describing-parameters/#path-parameters
* https://swagger.io/docs/specification/links/
* https://stackoverflow.com/a/45471010/1477964
* https://github.com/jsonapi-serializer/jsonapi-serializer/issues/53