Add use case to add items to a safebox
This code applies the same logic than previous use cases. Refer to the
related commits for more details.

A token is used for granting access to the contents of the safebox.

A couple of domain services are used to add a new item to a given
safebox and finding a safebox item by ID. The idea here was always
interacting with the items through the associated safebox.

These unit tests are currently in green, mocking the repository received
in the constructor.

There is a bit of duplication on these tests, but it's something that I
consider it's worth it, as everything involved is explicitly defined
within any given example.

Once again, the next step will be creating integration tests for the
repository.

References:

* https://docs.rubocop.org/rubocop/cops_metrics.html#metricsabcsize
* https://docs.rubocop.org/rubocop/cops_metrics.html#metricsmethodlength