Add missing logic required to adding items to a safebox
All possible responses from the server are now covered.

The acceptance tests are now in green.

Anything related to the token it's exactly the same here than with the
endpoint aimed to list the items of a safebox. Refer to those commits
for more details.

The logic for the domain services used inside the use case has been
implemented here as well.