Allow to paginate safebox items
Pagination links are included in the response of the endpoint returning
the list of items included in a safebox.

The parameterizable query now returns a hash containing the results of
the query and the number of total results found by that same query
without pagination nor filtering (if implemented).

In case an invalid parameter is provided, a bad request (400) is
returned, as stated in JSON:API specification.

The cop that prevents multiple example in a single test example has been
disabled for some files because I prefer to avoid lots of duplication in
those tests.

References:

* https://jsonapi.org/format/#fetching-pagination
* https://jsonapi.org/format/#query-parameters-families
* https://github.com/cerebris/jsonapi-resources/blob/master/lib/jsonapi/paginator.rb
* https://apidock.com/rails/v7.1.3.2/ActiveRecord/Calculations/count
* https://docs.rubocop.org/rubocop-rspec/cops_rspec.html#rspecmultipleexpectations
* https://stackoverflow.com/a/46475776/1477964